Create Install requests
In Python, using a proxy is usually to pretend to be another user's browser or server when sending HTTP requests. This can be achieved by using the requests library combined with the proxy server address. Here is a basic example using Python and a proxy:

First, make sure you have the requests library installed. If not, you can install it using pip: